requirements.txt
feedparser
supabase
python-dotenv
requests
beautifulsoup4